requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
transliterate==1.10.2
python-dotenv==1.0.1